+++ graphql_client_codegen/src/deprecation.rs
@@ -4,16 +4,23 @@ use syn;
 
 static DEPRECATION_ERROR: &'static str = "deprecated must be one of 'allow', 'deny', or 'warn'";
 
+/// Whether an item is deprecated, with context.
 #[derive(Debug, PartialEq, Hash, Clone)]
 pub enum DeprecationStatus {
+    /// Not deprecated
     Current,
+    /// Deprecated
     Deprecated(Option<String>),
 }
 
+/// The available deprecation startegies.
 #[derive(Debug, PartialEq)]
 pub enum DeprecationStrategy {
+    /// Allow use of deprecated items in queries, and say nothing.
     Allow,
+    /// Fail compilation if a deprecated item is used.
     Deny,
+    /// Allow use of deprecated items in queries, but warn about them (default).
     Warn,
 }
 
@@ -23,6 +30,7 @@ impl Default for DeprecationStrategy {
     }
 }
 
+/// Get the deprecation from a struct attribute in the derive case.
 pub fn extract_deprecation_strategy(
     ast: &syn::DeriveInput,
 ) -> Result<DeprecationStrategy, failure::Error> {