+++ graphql_client_codegen/src/fragments.rs
@@ -3,15 +3,21 @@ use query::QueryContext;
 use selection::Selection;
 use std::cell::Cell;
 
+/// Represents a fragment extracted from a query document.
 #[derive(Debug, PartialEq)]
 pub struct GqlFragment {
+    /// The name of the fragment, matching one-to-one with the name in the GraphQL query document.
     pub name: String,
+    /// The `on` clause of the fragment.
     pub on: String,
+    /// The selected fields.
     pub selection: Selection,
+    /// Whether the fragment
     pub is_required: Cell<bool>,
 }
 
 impl GqlFragment {
+    /// Generate all the Rust code required by the fragment's selection.
     pub(crate) fn to_rust(&self, context: &QueryContext) -> Result<TokenStream, ::failure::Error> {
         let derives = context.response_derives();
         let name_ident = Ident::new(&self.name, Span::call_site());